Modified the hull to be convex. Added trilinear voxel interpolation. Removed boundary refinement. Removed a generate triangle check. Removed a temporary face array in mesh projection. Added -preset slow to video encoding, removed maxrate limit, and added some info comments.

+++ render/Render/Volume.h
@@ -8,9 +8,9 @@ class Volume
 {
 public:
 
-	size_t Nx;
-	size_t Ny;
-	size_t Nz;
+	int Nx;
+	int Ny;
+	int Nz;
 
 	std::vector<float> Values;
